+++ Modules/Networking/networking_http_server.cpp
@@ -617,7 +617,7 @@ int32_t ft_http_server::run_once_locked()
         if (is_post && !body.empty())
         {
             std::snprintf(buffer, sizeof(buffer), "%llu",
-                static_cast<unsigned long long>(body.size()));
+                body.size());
             response.append("HTTP/1.1 200 OK\r\nContent-Length: ");
             response.append(buffer);
             if (should_keep_alive != FT_FALSE)
